Add integration tests to CI
This adds some simple tests for tab completions using an expect
script that tests the `example/example.cr` program interactively.

More tests should be added in the future for history and the
wrapper modules. It might also be possible to test hints but
I'm not sure since that data doesn't get echoed automatically.

Either way this is a good start.