Don't run tests on CI
It looks like the expect script hangs on CI for some reason. The
only thing I've been able to find online suggested that the github
actions runners use a dumb terminal which affects these types of
interactive scripts.

Adding tests is helpful regardless but this is annoying.

I also changed the version of the c compiler we use to build the
Linenoise library to see if that causes fewer warnings when building
on Linux.