Prefer shorter matches v2
This time we hopefully won't see the bugs we saw before which caused
the previous version to get reverted. It adds an option to sort
completions by string size instead of how they're normally sorted which
is alphabetically.

Other than that I added some tests for the Linenoise::Completion module
and changed the file name from completions.cr to completion.cr. This
shouldn't affect anyone using the library because this file is not
included directly. The new tests will run on CI unlike the interactive
expect tests which only get run locally for now.