Fix single line cursor offset bug
The cursor based its offset on the literal length of the line not
the visual length of the line which caused problems when using
escape codes. This fix was pulled in from an upstream branch and
it fixes the calculation for the cursor.

Note: This bug is still present in the multiline version though.

- https://github.com/antirez/linenoise/pull/135
- https://github.com/olegat