Merge in support for escape sequences and the ground work for UTF8
This is the first batch of changes from an upstream PR that provides
UTF8 support. As a side effect, it means that we now will have the
correct support for escape sequences in single and multiline mode.

We previously merged in a smaller change which provided support
for single line mode with escape sequences by fixing the cursor
position but this is more comprehensive.

The next commit will add the UTF8 library.

Author: https://github.com/yhirose
PR: https://github.com/antirez/linenoise/pull/187